refactor(lotto): change printing method
Change printing method
- previous method was to use StringBuilder
- test case requires System.out.println() for each sentence
- therefore removed StringBuilder

+++ src/main/java/lotto/controller/LottoResultController.java
@@ -17,7 +17,7 @@ public class LottoResultController {
     private List<LottoGrade> lottoGrades;
     private double yield;
 
-    private final int price;
+    private final double price;
     private int totalWinning;
 
     public LottoResultController(LottoAnswer lottoAnswer, List<Lotto> lottos, int price) {
